ping-infra-spring-boot-starter/src/test/java/cn/procsl/ping/boot/infra/web/user/UserControllerTest.java
package cn.procsl.ping.boot.infra.web.user;

import cn.procsl.ping.boot.infra.InfraApplication;
import com.fasterxml.jackson.databind.json.JsonMapper;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

@AutoConfigureMockMvc
@SpringBootTest(classes = InfraApplication.class, webEnvironment = SpringBootTest.WebEnvironment.MOCK)
public class UserControllerTest {

    @Autowired
    MockMvc mockMvc;

    @Autowired
    UserController userController;

    JsonMapper jsonMapper = new JsonMapper();

    @Test
    void test() {
        Assertions.assertNotNull(userController);
    }

    @Test
    public void register() throws Exception {
        UserDTO user = new UserDTO("program_chen@foxmail.com", "password");
        mockMvc.perform(
                        MockMvcRequestBuilders.post("/users")
                                .contentType(MediaType.APPLICATION_JSON)
                                .content(jsonMapper.writeValueAsBytes(user))
                )
                .andExpect(MockMvcResultMatchers.status().is2xxSuccessful())
                .andExpect(MockMvcResultMatchers.content().contentType("application/json"))
                .andDo(result -> {
                });
    }

}